UI: separate sensitivity plots, limit rerunnable ops
