+++ app/javascript/sensitivity_plotter/components/ScreeningScatterPlot.jsx
